extracted common methods
